HUGE Level Changes and Fine Tuning
Multiple things were edited and added or deleted. Firstly the UI layout and sheets became redundant and had to be deleted.

Player movements besides teleportation have been fine tuned to reduce sliding and quality pacing.

Level 2 and 3 were switched. Level 3 and 4 as well as an endgame screen were created.

What needs to be done now is for the mob team to add in the mobs as indicated, add audio, update sprite art assets.